# Conflicts:
#	src/search/Histogram.java
#	src/sort/Bound.java

+++ src/sandbox/Point.java
@@ -7,19 +7,19 @@ public class Point implements Comparable<Point>, Serializable {
 	 * 
 	 */
 	private static final long serialVersionUID = 5361956730616676054L;
-	private final int x;
-	private final int y;
+	private final double x;
+	private final double y;
 	
-	public Point(int x, int y) {
+	public Point(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
 	
-	public int getX() {
+	public double getX() {
 		return this.x;
 	}
 	
-	public int getY() {
+	public double getY() {
 		return this.y;
 	}
